image-gif-creator/requirements.txt
tkinter
Pillow
imageio
opencv-python